+++ resources/assets/js/store/modules/inventory/mutations.js
@@ -19,7 +19,7 @@ export default {
         const existing = filterResponse.map(ss => ss.name).indexOf(each.name)
         if(-1 < existing) {
           filterResponse[existing].quantity = parseFloat(filterResponse[existing].quantity) + parseFloat(each.quantity)
-          filterResponse[existing].price = (parseInt(filterResponse[existing].price) + parseInt(each.price))/2
+          filterResponse[existing].price = (parseFloat(filterResponse[existing].price) + parseFloat(each.price)) / 2
         } else {
           filterResponse.push(each)
         }
